requirements.txt
asgiref==3.3.4
autopep8==1.5.7
Django==3.2.4
django-ckeditor==6.1.0
django-debug-toolbar==3.2.1
django-js-asset==1.2.2
Pillow==8.2.0
pycodestyle==2.7.0
pytz==2021.1
sqlparse==0.4.1
toml==0.10.2
Unidecode==1.2.0
